05.01

+++ backend/app/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, DateTime, Index, Integer, JSON, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, DateTime, Index, Integer, JSON, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -39,7 +39,7 @@ class Dataset(Base):
     project_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
     artifact_id: Mapped[int | None] = mapped_column(Integer, nullable=True, index=True)
     content_type: Mapped[str] = mapped_column(String(255), nullable=False, default="text/csv")
-    size_bytes: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    size_bytes: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
     checksum_sha256: Mapped[str | None] = mapped_column(String(64), nullable=True, index=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=utcnow_naive, nullable=False)
